requirements.txt
diffusers[torch]
pillow
transformers
tkinterdnd2
tqdm
sv-ttk
